+++ CHANGELOG.md
@@ -6,10 +6,20 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html)
 and [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/).
 
+## [3.2.1] - 2025-08-24
+
+### Changed
+- Correct handling of "size" option.
+
+## [3.2.0] - 2025-08-24
+
+### Changed
+- Use much less memory when counting n-grams.
+
 ## [3.1.1] - 2025-08-24
 
 ### Changed
-- Implemented maximum n-gram  size.
+- Implemented maximum n-gram size.
 - Simplified handling of n-gram counter.
 
 ## [3.1.0] - 2025-08-23